src/requirements.txt
click
bibtexparser --pre
arxiv
python-dateutil
